React-Router 101
sec. 9, video 77

https://reacttraining.com/react-router/web/guides/quick-start

Install react-router-dom
$ npm i react-router-dom@4.2.2

We are going to use BrowserRouter once to create the new router and use Route for every single page.
Route is going to take 2 main things:
path: what URL do we want to use for this route.
component: when we match that URL what should we show to the screen.

Note that the API for BrowserRouter expects the child to either not exist or to have a length of 1.
So we create a <div> and put our routes into it.

Note that when React-Router matches our path, it just cares if the path at least starts with whatever we have.
We can overcome this with exact prop (default value is false).
https://reacttraining.com/react-router/web/api/Route/exact-bool

Originally, it was using the server side routing. e.g. If we visit localhost:8080/create, we would get 404 (Not Found) response.
What we need to do is make a tweak to our dev-server configuration. We need to send back index.html for all routes.
No matter which route we go to, we want to serve up index.html and let React-Router figure out what should get shown to the screen.
In order to do that, we need to modify webpack.config.js.
This modification tells the devServer that we will handle routing via our client side code and that it should return this page for all 404 routes.
https://webpack-v3.jsx.app/configuration/dev-server/#devserver-historyapifallback