Setting up a 404
sec. 9, video 78

https://reacttraining.com/react-router/web/api/Switch
https://reacttraining.com/react-router/web/api/Route/path-string-string

When React-Router sees Switch, it is going to move through our route definitions in order and it's going to stop when it finds a match.
If none of these match, switches are going to continue looking until it gets to the bottom. Note that Routes without a path always match.
So the NotFoundPage will show up for routes that are not defined.