Linking between routes
sec. 9, video 79

When we switch pages by manually changing the URL, we go through the full page refresh.
We are communicating with the server.
We want to be able to switch pages without going through the full page refresh.

We will add an event listener for our links then we're going to override the browser's default behavior.
https://reacttraining.com/react-router/web/api/Link
https://reacttraining.com/react-router/web/api/NavLink

With Link, we are using client side routing as opposed to server side routing behind the scenes.
So when we are linking internally in our application,
we want to make sure to use a <Link> as opposed to <a> tag in order to take advantage of client side routing.

<NavLink> is a special version of the <Link> that will add styling attributes to the rendered element when it matches the current URL,
which is better suited for situations like navigation.

activeClassName is going to allow us to provide a class.
This class is only going to get applied to the link when we're on that page.
That's going to allow us to target the current active link.

Note that NavLink just cares if the path at least starts with whatever we have.
We need to use exact attribute to fix it.