Query strings and URL parameters
sec. 9, video 81

When React-Router finds a route and it's a match, not only is it rendering the component, it's also passing a few props down.
So any time we use a component inside of a route, we get access to some special info.

It's worth noting that React-Router is only going to pass these props down to the components that are actually used inside of a <Route>.
In our case Header component is not going to have access because Header was not set up as the component value of a <Route>.