Setting up Redux
sec. 10, video 85

We're going to explore Redux in isolation.
We're going to have webpack run redux-101.js in the playground folder instead of app.js.

https://redux.js.org/
https://devdocs.io/redux/

Install Redux
$ npm i redux@3.7.2

We create a Redux store (by calling createStore) that holds the complete state tree of our app.
There should be only a single store in our app.
createStore expects a reducer to be the 1st argument.
A reducer, a pure function with (state, action) => state signature.
It describes how an action transforms the state into the next state.
We also set the default state value in the reducer.
The important part is that we should not mutate the state object, but return a new object if the state changes.

store.getState returns the current state object.