Dispatching Actions
sec. 10, video 86

The only way to mutate the internal state is to dispatch an action.
Action is an object that gets sent to the store and describes the type of action we'd like to take.

Note that when we call store.dispatch(action), the reducer (the 1st argument of createStore) runs again.
We can access the passed in action through the 2nd argument of reducer.
Note that we should not mutate the state object, but return a new object if the state changes.