Subscribing and dynamic actions
sec. 10, video 87

We can pass a single function to store.subscribe() and this function gets called every single time the store changes.
What's more, we can get a returned function after calling store.subscribe().
It can be used to remove this change listener.

Note that type prop has to be provided for the action object.
If we don't provide a type prop, Redux is going to throw an error.

We can pass dynamic information along inside of our action objects.
The store can then use that information to calculate the new state.