Refactoring and organizing
sec. 10, video 90

 https://redux.js.org/basics/actions

Our target is to create Action generators which return action objects.
Originally, we use ternary operator to check if the passed in parameters satisfy some conditions.
Now we can use object destructuring and default value setting to simplify our code:

({ incrementBy = 1 } = {}) => ({ … }) means if there is no passed in parameter, the default parameter is an empty object.
Then we destructured the incrementBy prop from the passed in object.
If there is no incrementBy prop, the default value would be 1.
Finally, we use the object property shorthand syntax.