Filtering redux data
sec. 10, video 96

Our target is that we only show the expenses which match the text, startDate and endDate in filters.
Note that if startDate was not provided (the value would be undefined), the startDateMatch would be true.
If startDate was not provided as a number, it will check if expense.created is greater than or equal to startDate.