The higher order component
sec. 11, video 100

Higher Order Component (HOC) - A component (HOC) that renders another component.
The advantages of using HOC:
	1. Reuse code
	2. Render hijacking
	3. Prop manipulation
	4. Abstract state

Note that when we instantiating a component inside of JSX, we can actually create a JavaScript expression.
And we can spread out any given object we like.
This has the effect of taking every key value pair on that object and passing them down as props.

In this case, we create a regular function that returns the higher order component.