Connect Store and Component with React-Redux
sec. 11, video 101

https://react-redux.js.org/introduction/quick-start

Install react-redux
$ npm i react-redux@5.0.5

React-Redux provides <Provider />, which makes the Redux store available to the rest of our app.
The connect() function connects a React component to a Redux store.
The returned value (a function) of connect() is just like the function we create in the last video which takes our component and returns the higher order component.
mapStateToProps, the function we pass into connect() as the 1st argument, lets us determine what information from the store we want our component to be able to access.
And the store's state actually gets passed in.
The results of mapStateToProps must be a plain object, which will be merged into the component’s props.
The new component will subscribe to Redux store updates.
This means that any time the store is updated, mapStateToProps will be called, getting the fresh values in the component.