Controlled inputs for filters
sec. 11, video 103

Our 1st target is create an <input /> so users can type the filter text.
When users type in the filter text, it will dispatch a 'SET_TEXT_FILTER'  action to change the state.

When we connect a component to redux, we actually get some other information passed into the component.

Through React dev tool, we can see that we have dispatch prop, which is the exact same one that we were accessing on the store.
So we have access to dispatch from inside of our connected components.

The 2nd target of this video is add a button for each expense.
When the button is clicked, it will dispatch a 'REMOVE_EXPENSE' action.

Originally, <ExpenseListItem /> was not connected to redux.
The props were passed from <ExpenseList />. However, we need to dispatch an action.
Note that we can just inject dispatch into the component without listening to store if we don't need anything from the state.
This can be done by not providing any parameter to connect().
This is going to give us access to the dispatch prop.

https://react-redux.js.org/5.x/api/connect#inject-just-dispatch-and-don-t-listen-to-store