Setting up a date picker
sec. 11, video 106

https://momentjs.com/
https://github.com/airbnb/react-dates
https://github.com/airbnb/react-dates#singledatepicker

Install moment & react-dates
$ npm i moment@2.18.1 react-dates@12.7.0

What's more, react-dates does require another peer dependency that we have to install.
We're not going to be using it directly but it's used internally by the react-dates library.
$ npm i react-addons-shallow-compare@15.6.0

Note that we need to import _datepicker.css as well.
https://github.com/airbnb/react-dates#webpack

For <SingleDatePicker/>, the date prop is just a moment object that represents where you want to start.
We configure the number of months that were shown (numberOfMonths) to 1.
Note that by default the <SingleDatePicker/> only lets us pick the current day or days in the future.
We can override isOutsideRange to fix it. We provide a function always returns false, which means that every single day is available to us.