Wiring up Add Expense
sec. 11, video 107

We are going to setup onSubmit for the form and that's going to allow us to dispatch the action that adds the expense and
then will redirect the user over to the dashboard page.

Because we want to reuse the <ExpenseForm />, so we do not dispatch the action inside of <ExpenseForm />.
Instead, we pass a prop from <AddExpensePage />, which is a function named onSubmit.
<ExpenseForm /> will send the expense data to its parent component through this function and it will actually dispatch the given action to the redux store.
Which means we need to connect <AddExpensePage /> to redux store.

https://momentjs.com/docs/#/displaying/unix-timestamp-milliseconds/

To automatically redirect to dashboard page, we can use the push method in history, which is a prop passed in by <Route />.
Ref: Query strings and URL parameters
Push is how we can programmatically change pages. It's going to switch over using browser routing.