Wiring up Edit Expense
sec. 11, video 108

First, we add a <Link /> to our <ExpenseListItem /> which will redirect us to <EditExpensePage />.
We no longer remove the expense in <ExpenseListItem /> anymore, so we remove the connect import from it.

Then we connect <EditExpensePage /> to redux store. We want to be able to give the component the current expense object.

https://react-redux.js.org/using-react-redux/connect-mapstate#arguments

If our component needs the data from its own props to retrieve data from the store,
we can define the mapStateToProps function with a 2nd argument, ownProps.

When we try to edit the expense, we need to see the current values for the selected expense.
So we need to modify <ExpenseForm /> to only show the default values if no expense was passed down.
And we need to pass the props into <ExpenseForm /> for editing an expense.
For adding a new expense, we will not pass in an expense and the default values will be shown on <ExpenseForm />.
Because we need to access the passed in props, we need to define our state in the constructor function.

Then after we click the button (Add Expense), we need to dispatch the 'EDIT_EXPENSE' action and redirect to the dashboard page.
Finally, we add a <button /> to remove the expense in <EditExpensePage />.