Filtering by dates
sec. 11, video 110

https://momentjs.com/docs/#/manipulating/start-of/

First, we modify the default values of startDate & endDate of filters state to the beginning and the end date of the current month.

https://github.com/airbnb/react-dates#daterangepicker

Then we setup <DateRangePicker /> in <ExpenseListFilters />.
Note that in <ExpenseForm />, we track all the props of <SingleDatePicker /> in local state.
However, for <ExpenseListFilters />, because we have the filters related props in redux store, we dispatch actions to change the props values.
Because we still need to track focusedInput prop of <DateRangePicker /> as local state, we need to switch <ExpenseListFilters /> to class based component.
What's more, we enable the ability to clear the dates with a little X button by setting the showClearDates to true.

https://momentjs.com/docs/#/parsing/unix-timestamp-milliseconds/
https://momentjs.com/docs/#/query/

Then we use moment to filter expense by dates.