Test expenses selector
sec. 12, video 115

https://momentjs.com/docs/#/manipulating/

To test expenses selector, we need to create an array of expense (test data).
To generate expenses with different createdAt days, we can use moment.add & moment.subtract.

For each test case, we are going to define our filters and pass it in selectedExpenses function, get the result back and then make the assertion.

https://jestjs.io/docs/en/22.x/expect#toequalvalue

Note that to assert an array, we can use .toEqual as well.