Testing filters reducer
sec. 12, video 116

The 1st test case is the one that makes sure the default values get set up correctly when the redux store first kicks off.
Redux dispatches a special action, type: '@@INIT', for that. This is used internally by redux.
We can see that through redux dev tool.

The key is that we need to fully understand what the function we want to test really does. What is the input and what is the expected output.
In this case, filtersReducer expects a current state (default value will be provided if it is undefined) and an action.
Then it returns a updated state. So we setup different test cases with undefined as current state and different type props and additional data in action object.

Because the default value of sortBy is 'date', so we need to provide a current state with sortBy as 'amount' to make sure that it actually being updated.