Snapshot testing
sec. 12, video 118

To test our components, we need to figure out how we're going to virtually render our component.
We need a way to figure out what JSX comes back.
But we're not going to be viewing it in the browser.
React-test-renderer allows us to render our components inside of just regular JavaScript code and then we can assert something about what go rendered.
There are actually 2 main ways we can test react components: Shallow rendering and full DOM rendering.

https://reactjs.org/docs/shallow-renderer.html

$ npm i react-test-renderer@16.0.0

https://jestjs.io/docs/en/22.x/expect#tomatchsnapshotoptionalstring

We are not going to use .toEqual and try to define the expected object.
Snapshots allow us to track changes to data over time.
So we are going to able to create a snapshot of <Header /> at its current point in time and we're going to be able to get notified if this ever changes.
When using .toMatchSnapshot, the 1st time we run this test case, it's always going to pass because there is no existing snapshot.
So jest is going to create a new one. It's going to create a snapshot of what the rendered <Header /> output looked like.
The 2nd time we run this test case it's going to compare with the existing one. If it's the same, the test will pass. Otherwise, the test is going to fail.

When we run the test, a __snapshot__ folder will be generated by jest (we should not change any of the files inside).

If we change something accidentally and rerun the tests, we will get an error.

If we actually want to update our component, e.g. we want to remove the <NavLink /> to help page from <Header />.
We can follow the hint shown in the following error message.
So if we press the 'u' key and it's going to take the new snapshot and replace the old one.