Enzyme
sec. 12, video 119

https://airbnb.io/enzyme/
https://www.npmjs.com/package/raf

We are going to replace react-test-renderer with Enzyme.
Enzyme is a renderer for react but is a much more full featured renderer.
What's more, we need to install the adapter which allows us to specify exactly which version of react we want to test against.
This allows the core enzyme library to be much smaller.
We also need to install a polyfill for a browser feature known as request animation frame.
This is something provided by our browsers. And when we don't have it in this testing environment, it does cause some issues.

$ npm i enzyme@3.0.0 enzyme-adapter-react-16@1.0.0 raf@3.3.2

Then we create a setup test file in our project. It allows us to configure our test environment.
This is where we're going to be able to set up the enzyme adapter. So we create setupTests.js in tests folder.

https://jestjs.io/docs/en/22.x/configuration#setupfiles-array
https://www.npmjs.com/package/raf#rafpolyfillobject

The next thing we're going to do is setup a jest configuration json file, which allows us to actually specify that the setup test file should run.
Note that this file is going to live in the root of our project.

Then we have to change our test script in package.json.

https://airbnb.io/enzyme/docs/api/shallow.html

Then we can use shallow rendering API of enzyme.
Here, we can use .find(selector) to select the various elements inside of our component and make assertions about them.
e.g. we can assert that there is 1 <h1> tag in <Header />.
We can also make an assertion about the text value of that <h1>.

The goal is actually create a snapshot based off of the enzyme wrapper.
To make enzyme work with snapshot testing functionality, there is 1 utility library we have to install.

https://www.npmjs.com/package/enzyme-to-json
$ npm i enzyme-to-json@3.0.1

toJSON is going to take the wrapper and extract just the meaningful stuff, the rendered output
(convert any Enzyme wrapper to a format compatible with Jest snapshot).

Then we can setup enzyme-to-json to work automatically so we don't have to add it into our code.

https://github.com/adriantoine/enzyme-to-json/tree/f042dbcf8ce8fc0ef8ac121f9548ffe3edb6e71b#usage

Finally we can remove the toJSON code from Header.test.js