Snapshot testing with dynamic components
sec. 12, video 120

When we're testing our react components, we actually want to test the unconnected version because we want to be able to provide a set of dynamic props.
So we don't actually want these props come from the store. Instead, we're just going to provide expenses directly.
So we have to export the unconnected component (we still export the connected version by default).