Mocking libraries with Jest
sec. 12, video 121

First, we move the importing of _datepicker.css from ExpenseForm.js to app.js. (we will manage these in the next section)

To test <ExpenseForm />, note that the createdAt prop always uses the current point in time.
It will cause the failure of snapshot testing.
We need a way to consistently get the same data back without changing it.

To fix this, we are going to mock out moment.
We will create a fake version of the moment library that is going to allow us to define what happens when the code actually calls moment().

https://jestjs.io/docs/en/22.x/manual-mocks

First, we create moment.js in __mocks__ folder in tests folder.
To ensure that a manual mock and its real implementation stay in sync,
it might be useful to require the real module using jest.requireActual(moduleName) in your manual mock and amending it with mock functions before exporting it.
We force the moment to start a specific point in time if no point in time was provided.