Testing user interaction
sec. 12, video 122

https://airbnb.io/enzyme/docs/api/ShallowWrapper/simulate.html
https://airbnb.io/enzyme/docs/api/ShallowWrapper/state.html

Here, we are going to simulate the form submission.
Note that when we submit the form, we first run event.preventDefault().
So we need to pass in the event object which contains a method named preventDefault.
Then we can use the state method to fetch the state for that component.
What's more, we can add another snapshot beforehand to make sure that no errors show up (we can add as many snapshots as we like into our test cases).

https://airbnb.io/enzyme/docs/api/ShallowWrapper/at.html

Then we can create a test case that change the <input/> value for description and assert the state value.
Because we have multiple <input/> tags in the form, we can use .at(0) to match the 1st one.