Testing spies
sec. 12, video 123

https://jestjs.io/docs/en/22.x/jest-object#jestfnimplementation
https://jestjs.io/docs/en/22.x/mock-function-api

The goal of spies is to create functions that are fake functions.
They're created by Jest for us and we can make assertions about them.

We are able to pass the spy into the component that we render,
simulate things like the form submission and actually ensure that everything went well.

https://airbnb.io/enzyme/docs/api/ShallowWrapper/prop.html

Then we want to test onDateChange method for <SingleDatePicker />.
So we need to get this method and call it.
Now we can make sure that when <SingleDatePicker /> does fire onDateChange, our component is acting correctly.
Here, we can use .prop to get the prop value (which is a function in our case).

We can use the same technique to test onFocusChange of <SingleDatePicker />.