Testing AddExpensePage
sec. 12, video 124

Originally, we use props.dispatch(addExpense(expense)) in <AddExpensePage/>.
Because addExpense is imported from other file, not something that's passed in as a prop.
It's a little harder for us to test.

https://react-redux.js.org/5.x/using-react-redux/connect-mapdispatch

connect can accept an argument called mapDispatchToProps, which lets us create functions that dispatch when called, and pass those functions as props to our component.
It is a way to return our dispatcher functions allowing us to abstract them away from the component itself.
Once you've wrapped all our action creators with functions that dispatch the actions, the component is free of the need of dispatch.
Therefore, if you define your own mapDispatchToProps, the connected component will no longer receive dispatch.

To avoid inline functions, we switch <AddExpensePage/> to a class based component.
Because we want to test the unconnected version of component, we need to export it.

Now we can create AddExpensePage.test.js and create some test cases.
We can create 2 spies for onSubmit and history.push and assert that when the onSubmit method of <AddExpensePage/> was called, they were actually called with correct arguments.

https://jestjs.io/docs/en/22.x/api

Note that the above 2 test cases have the same 3 lines in the beginning of the test.
So we can use beforeEach to set up the spies and the component.