Testing EditExpensePage
sec. 12, video 125

We can use the same technique which was applied on testing <AddExpensePage/> to test <EditExpensePage/>.
First, we switch <EditExpensePage/> to class based component.
Then we can create EditExpensePage.test.js.
Note that we need to pass expense into <EditExpensePage/> as a prop because it receives the expense prop through mapStateToProps.