Production Webpack
https://webpack-v3.jsx.app/guides/production/#cli-alternatives
https://webpack-v3.jsx.app/api/cli/#environment-options

First, we create a new build script of webpack for production mode in package.json.
Using -p flag allows us to minimize our JavaScript and sets the production environment variable.
We also set up the --env flag, which allows us to customize webpack.config.js (which will be explored in the following).

https://webpack-v3.jsx.app/configuration/configuration-types/#exporting-multiple-configurations

2nd, we change the configuration type of webpack.config.js to exporting a function.
The advantage is that the function actually gets called with env.

https://webpack-v3.jsx.app/configuration/devtool/

Note that the source-maps takes the most part of the size of bundle.js.
We will optimize it for a slower version (source-map).
source-map takes a lot more time to build and that is an external file which is great for production.

The external file (bundle.js.map) is only going to get loaded when someone actually cracks open the developer tools.
For regular users, the browser is never going to make a request for this file.