Creating separate CSS files
sec. 13, video 134

Setup production webpack build.

Currently, all of css settings live inside of bundle.js.
This means that bundle.js is bigger than it need to be and it's also doing things that it shouldn't be doing.
When we have our styles in bundle.js, they don't actually get added to the browser until after the JavaScript runs, which takes some time.
So what we're going to do is have webpack output a JavaScript file which contains just JavaScript and a separate css file which contains all of our styles.
We can then link that into index.html.

https://www.npmjs.com/package/extract-text-webpack-plugin
Install extract-text-webpack-plugin
$ npm i extract-text-webpack-plugin@3.0.0

We have essentially told webpack to "when you see stuff extract it". We’re still going to process it.
Note that we're no longer going to use style-loader because style-loader handled the actual in-lining of the styles.
Then we add a <link/> tag in index.html.

https://github.com/webpack-contrib/css-loader#sourcemap
Finally, we setup source-map for css files.