A production web server with Express
sec. 13, video 135

Install express
$ npm i express@4.15.4

Here we create an express server and set the path to the folder for static files.
Note that we add one more get router to match every request path.
So if we refresh the page one the path /create, the server would still send the index.html back.