Adding total selector
sec. 13, video 139

Our target is add a function that takes an array of expenses and returns the total amount.
So we create expenses-total.js in selectors folder and the corresponding expenses-total.test.js.

For test driven development (TDD), we write the test case first and then we make a tweak to the code to get that test case to pass.