Getting Firebase
sec. 14, video 142

Create a new project (Expensify) on Firebase.
Create a database with development mode.
Add firebase to our web app.

We need to create a nickname for our app first.
After that, we can have our firebaseConfig object.

Install firebase (https://www.npmjs.com/package/firebase)
$ npm i firebase@4.2.0

Create a new folder (firebase) in src folder.
Create firebase.js in the folder. This is where we are going to connect to the database.
So we're  going to connect to the database once in this file and then other files in our project can  use the connection by just importing what we have here.

Note that we take all of the named exports from 'firebase' and dump them on a brand new variable called firebase.

Paste the firebaseConfig object up on the firebase.js. And we can test our connection.
Note that we need to import firebase.js in app.js to get the file run.

Run the dev-server
$ npm run dev-server

Then we can check the database on Firebase.