Testing async redux actions: part I
sec. 14, video 153

First, we update AddExpensePage.test.js because the component calls startAddExpense instead of addExpense.

We're going to create a fake redux store for testing purposes.
To test startAddExpense, what we care about are 2 things, the database was successfully updated and the correct action was dispatched.

Install redux-mock-store (https://github.com/dmitry-zaets/redux-mock-store)
$ npm i redux-mock-store@1.2.3 --save-dev

Then we can initialize mockstore with an empty object inside of our test case.
Note that if we want to force Jest to wait until a specific point in time, we have to provide 'done'.
And this test case is no longer going to be considered a success of a failure until after we call done.

https://github.com/dmitry-zaets/redux-mock-store#asynchronous-actions
Because we want to wait for the firebase call to actually finish and then run some assertions, we use the Promise chaining.
We need to return a Promise (which sends the push() command to Firebase and dispatches addExpense action) so we can use the then() call.