Testing async redux actions: part II
sec. 14, video 154

Now, we can get the actions that were dispatched to the mock store with store.getActions() and make some assertion about them.
Then we can fetch data from Firebase with once() and see if it was actually saved over there.
Note that we return a Promise here so we can use Promise chaining instead of nested callbacks.