Creating a separate test database
sec. 14, video 155

To setup environment variables regardless of operating system,
lets install cross-env (https://www.npmjs.com/package/cross-env)
$ npm i cross-env@5.0.5 --save-dev

We only need to change to the script for test.
There is no need to do it for production since it's set up for us by Heroku.

Create .env.dev in the root of our project and paste the firebaseConfig into it.

Create a new project (Expensify-test) on Firebase for test purpose.
Create a database in test mode and add Firebase to our web app.

Create .env.test in the root of our project and paste the firebaseConfig for test into it.

Install dotenv (https://www.npmjs.com/package/dotenv) to load environment variables from a .env file into process.env.
$ npm i dotenv@4.0.0 --save-dev

https://www.npmjs.com/package/dotenv#path

Setup dotenv in webpack.config.js.

https://webpack.js.org/plugins/define-plugin/

Note that our Node environment variables, the ones that exist in the webpack.config.js,
they do not get passed down to the client side JavaScript.
So we need to manually pass those through.

Using dotenv in setupTest.js in tests folder.

Now, when we re-run the test, it will access database for test, not the original one.