Heroku environment variables
sec. 14, video 156

To setup the environment variables used in firebase.js on Heroku, we will use Heroku CLI.

To check the environment variables that were set:
$ heroku config

To set the environment variables
$ heroku config:set KEY=value

To remove the environment variables
$ heroku config:unset KEY

Now, we can set the environment variables for production.

Add the env file into .gitignore