Fetching expenses: part II
sec. 14, video 158

Create startSetExpenses action generator to fetch data from Firebase and dispatch 'SET_EXPENSES' action.
Then we add a test case to make sure the correct action was dispatched.
In app.js, we import startSetExpenses and dispatch it to get the data from Firebase.
Finally, we can push the commit on to GitHub & Heroku.