Login page and Google authentication
sec. 16, video 162

https://firebase.google.com/docs/auth/web/google-signin?authuser=0
https://firebase.google.com/docs/reference/js/firebase.auth.GoogleAuthProvider?authuser=0

Setup Google authentication on Firebase.

In firebase.js, we export an instance of the Google provider object.

In actions folder, create auth.js.
Create an action generator which returns a function that signs in with pop-up window.

Create <LoginPage/> to let user sign in with Google authentication.
Add LoginPage.test.js to check the snapshot.

Add <LoginPage/> to AppRouter.js and render it in the root path.

In app.js, we can use onAuthStateChanged() to check if we sign in successfully (temporarily).