Redirecting Login or Logout
sec. 16, video 164

Originally, the history prop is passed into the component, e.g. <AddExpensePage/> because it is registered with a <Route>.
But the problem is if we are not inside of a component registered to a <Route>, we couldn't access that history API.
To be able to use the history API anywhere, not just in the context of a component, we install history (https://www.npmjs.com/package/history).

$ npm i history@4.7.2

https://reacttraining.com/react-router/web/api/Router

In AppRouter.js, we were using <BrowserRouter>, which uses the browser history by default.
Behind the scenes React-router is creating an instance of browser history and registering it with our new Router.
We can go through that process manually. We are going to switch from <BrowserRouter> to the regular <Router>.
Then we are allowed to provide our own history value.

https://firebase.google.com/docs/auth/web/manage-users
https://firebase.google.com/docs/reference/android/com/google/firebase/auth/FirebaseAuth.AuthStateListener#public-methods_1

In app.js, if someone does log in, we will fetch their expenses. Otherwise, we will get the user to login page.
Note that we want to make sure the app only renders a single time.
What's more, if the user was on login page and he actually signs in, we will redirect him to the dashboard page.