The Auth reducer
sec. 16, video 165

https://firebase.google.com/docs/auth/web/manage-users

We're going to keep track of whether or not the user is logged in and we will do that by storing the user's ID (user.uid) somewhere in the Redux store.
We can then use that value throughout the application to figure out if we're logged in.
If we're not logged in, uid would be undefined. We create 2 action generators and the corresponding reducer for authentication.

Connect the reducer to the Redux store in configureStore.js.

Dispatch login/logout actions in the observer onAuthStateChanged().
The reason we dispatch login action in the observer as opposed to dispatching it somewhere inside the startLogin,
like we've done for our other async actions, is because the callback of the observer is also going to run when a user first visits the web page.
And we can make sure the Redux store is up-to-date.
If we only dispatched login/logout inside of our async actions, the value would only ever get set or cleared when someone explicitly logged in/out,
as opposed to when they're implicitly logged in by just loading the application after having been logged in before.

Write test cases for auth action generators and auth reducer.