Private only routes
sec. 16, video 166

In <AppRouter/>, we are going to modify how we set up the routes.
We're going to add a couple of new components that make sure to actually run the check for authentication before rendering the private components.
We will create a <PrivateRoute> in place of <Route> for private pages.

We can use <PrivateRoute/> and remove <Header/> in <AppRouter/>.

Create PrivateRoute.js in routers folder.
<PrivateRoute/> is a wrapper around <Route>.
The whole point of using it is to add some conditional logic in it which allows us to determine if the user is authenticated or not.
We can take the correct action either rendering the private stuff or redirecting them over to a public page.
When destructuring, because we're eventually going to be rendering the component and we know we need an uppercase name for that.
So we're going to rename it to uppercase Component.

https://developer.mozilla.org/zh-TW/docs/Web/JavaScript/Reference/Operators/Destructuring_assignment#Rest_in_Object_Destructuring

Here, we use the rest syntax for destructuring.
With …rest we get a variable called rest with all the stuff that we didn’t destructure.

https://reacttraining.com/react-router/web/api/Redirect

We also use <Redirect/> to navigate to a new location.
Note that we get all of the props that were passed to <Route> and pass those through to the individual component (Query strings and URL parameters).
…rest contains path or exact those passed to <PrivateRoute/>. props contains match, location, history… etc. those passed by <Route>.