Private Firebase data
sec. 16, video 168

We're going to have a private place for each user.

Our target data structure on Firebase is as following:

const db = {
    users: {
        $uid: {
            expenses: {
                asjodqwod: {

                },
                aswdjodwq: {

                }
            }
        }
    }
}

https://github.com/reduxjs/redux-thunk#motivation

When we have thunk actions, our async actions, we can call the 2nd argument getstate to get the current state.
Then we can get the uid. So we can update the path we access our users data on Firebase.
We apply the change to startAddExpense, startRemoveExpense, startEditExpense and startSetExpense.

For testing, because we will get the uid from state, we need to make sure that there is actually a uid set up in the store.
So when startRemoveExpense/startEditExpense actually runs, it has all of the necessary information.

https://firebase.google.com/docs/database/security/user-security

On Firebase, make ensure that each user can only write to a user-specific path.
So no one can read/write the data on root path.
Every user can only access the data under the path 'users/uid' if he is logged in.