Styling login page
sec. 17, video 171

Remember our partial files need to start with an underscore but the underscore and the extension are excluded when we're importing.

Note that when we put the path in the web server to the image we want, the code runs in the browser and the browser only has access to what's in the public directory.
By default the container is not going to take up the entire browser width or height. So we set it up to 100% viewport height/width.

https://www.w3schools.com/cssref/css3_pr_background-size.asp

We set the value of background-size to cover to always cover the container with the image.
So as we adjust the image or as we adjust the browser's height/width, we want to make sure it does still cove the entire box layout.

Now we want to get a box around the Login <button> and actually center that in the screen.

https://sass-lang.com/documentation/functions/color#transparentize

Here we use fade-out() to make color more transparent.

We apply the flex box (display: flex) to center the box.
We use just-content: center & align-items: center to have things centered in both directions.

https://www.w3schools.com/cssref/pr_text_text-align.asp

Using text-align: center to center the text in the box.

https://www.w3schools.com/cssref/pr_dim_line-height.asp

In our case with this title, we know it's always going to be setting on one line so we can adjust the line-height to 1,
allowing us to more evenly distribute the space above and below the <h1>.

Then we create the color for our body font.
Apply the styling to <LoginPage/>.