Styling buttons
sec. 17, video 172

Create _button.scss to style the login <button>.

Remove the <NavLink> for creating expense from <Header>.
Then Replace the <NavLink> for dashboard with <Link> because we just need a link to the dashboard instead of a navigation link with an active class.

For the title of the <Header>, we set the value of text-decoration to none,
which allows us to clear the default text decoration for links which is an underline.

Create _content-container.scss to make sure that we get everything centered appropriately.

The purpose of .header__content is to use flex box to distribute things (title & logout button) left to right.