Styling summary area
sec. 17, video 173

Create _page-header.scss to add some styles for <ExpensesSummary/>.

https://www.w3schools.com/css/css_inline-block.asp
For the <Link> which adds expense, we want to make the link look like a <button>.
Here. We set the value of display to inline-block to make sure that the vertical margin and padding actually get taken into effect.
We set the value of text-decoration to none to remove the underline from the link.

To make the logout <button> look like a link, we create a modifier (button--link) to set the background to none.

Note that we also use content-container to center the content of <ExpensesSummary/>.